bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["gavril-s <gavril.senkevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
common = {path = "../common", develop=true}
python-telegram-bot = "^20.8"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
